Add path test

+++ SimeonInstaller.ps1
@@ -1009,8 +1009,10 @@ CRLFOption=CRLFAlways
                 Write-Verbose "Deleting existing .git directory"
                 Remove-Item '.\.git' -Recurse -Force
 
-                Write-Verbose "Deleting existing .github directory"
-                Remove-Item '.\.github' -Recurse -Force
+                if (Test-Path -Path '.\.github') {
+                    Write-Verbose "Deleting existing .github directory"
+                    Remove-Item '.\.github' -Recurse -Force
+                }
 
                 Write-Verbose "Initializing new git repository with existing contents"
                 Invoke-CommandLine "git init 2>&1" | Write-Verbose
